test: move Parameter test mixin classes
